refactor: clean up command palette types
